# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Application API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the generated application",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "https://your-production-url.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "security": [],
  "paths": {
...
    "/api/audit-logs": {
      "get": {
        "summary": "Get audit logs",
        "tags": [
          "Audit"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Authenticate user",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "summary": "Get current user details",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User details retrieved"
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "summary": "Register a new user",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email",
                  "password",
                  "role"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully"
          }
        }
      }
    },
    "/api/contract-drafts/{id}": {
      "get": {
        "summary": "Get draft by ID",
        "tags": [
          "Drafts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "summary": "Update draft",
        "tags": [
          "Drafts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Delete draft",
        "tags": [
          "Drafts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/contract-drafts": {
      "get": {
        "summary": "Get all contract drafts",
        "tags": [
          "Drafts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Create new draft",
        "tags": [
          "Drafts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/contract-monitoring/{id}": {
      "get": {
        "summary": "Get monitoring record by ID",
        "tags": [
          "Monitoring"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/contract-monitoring": {
      "get": {
        "summary": "Get all monitoring records",
        "tags": [
          "Monitoring"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Create monitoring record",
        "tags": [
          "Monitoring"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
...
}